Se obtine la cantidad de comisiones por oferta académica

+++ src/main/java/ar/edu/unq/inscripcionunq/spring/controller/miniobject/OfertaAcademicaJson.java
@@ -7,7 +7,7 @@ public class OfertaAcademicaJson {
 	public String descripcion;
 	public boolean habilitada;
 	public CarreraJson carrera;
-	public int nroComisionesCreadas;
+	public Long nroComisionesCreadas;
     
 	
 
@@ -24,11 +24,11 @@ public class OfertaAcademicaJson {
 		super();
 	}
 
-	public int getNroComisionesCreadas() {
+	public Long getNroComisionesCreadas() {
 		return nroComisionesCreadas;
 	}
 
-	public void setNroComisionesCreadas(int nroComisionesCreadas) {
+	public void setNroComisionesCreadas(Long nroComisionesCreadas) {
 		this.nroComisionesCreadas = nroComisionesCreadas;
 	}
